Merged PR 5438: Update PPiOS-Rename to support Xcode 12 and iOS 14
The binary format of libraries changed somewhat. A new flag was added to a field indicating that three pointers describing methods would be a compressed relative format. A level of indirection was also added for the name fields (__objc_selrefs) so that these relative pointers pointed into a data region that is an array of pointers pointing to the actual location of the method names (selectors).

Related work items: #19525